code split into separate functions; minor changes
There is now a dedicated function for
a) checking for commas
b) finding a class match for the adder function
c) processing the classes into a string for the adder function

minor changes:
added some commands to the commands list
added a command command so you can view the commands by typing either
'command' or 'help command'. However, this function still is a WIP!
enjoy!